Move W3C trace context propagator to propagators
